Gets CSRF working with auth and proxies ember reloads
