+++ mysite/urls.py
@@ -26,6 +26,8 @@ from django.contrib.staticfiles import views
 from rest_framework.authtoken.views import obtain_auth_token
 from rest_framework.authentication import TokenAuthentication
 from rest_framework.permissions import IsAuthenticated
+from urllib.request import urlopen
+from django.http import HttpResponse
 
 class AuthorsSerializer(serializers.HyperlinkedModelSerializer):
     class Meta:
@@ -78,16 +80,22 @@ router.register(r'authors', AuthorViewSet)
 router.register(r'genres', GenreViewSet)
 
 urlpatterns = [
+    url(r'^api/api-auth/', include('rest_framework.urls')),
+    url(r'^api/api-auth-token', obtain_auth_token),
     url(r'^api/', include(router.urls)),
-    url(r'^api-auth-token', obtain_auth_token),
     url(r'^books/', include('books.urls')),
     url(r'^admin/', admin.site.urls),
-    url(r'^api-auth/', include('rest_framework.urls')),
 ]
+
+def proxy_live_reload(request):
+    url = "http://localhost:4200/ember-cli-live-reload.js"
+    response = urlopen(url)
+    return HttpResponse(response.read())
 
 if settings.DEBUG:
     urlpatterns += [
         url(r'^assets/(?P<path>.*)$', views.serve),
+        url(r'^ember-cli-live-reload.js$', proxy_live_reload),
     ]
 
 def ember (request):
